카카오 로그인 서비스 구현 완료

+++ src/main/java/com/cos/blog/repository/UserRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 // 자동으로 빈으로 등록된다.
 // @Repository 생략 가능
-public interface UserRepository extends JpaRepository<User,Integer> { // 해당 jpaRepository는 User테이블이 관리하는 레포징토리고 이 User테이블의 pk는 Integer다.
+public interface UserRepository extends JpaRepository<User,Long> { // 해당 jpaRepository는 User테이블이 관리하는 레포징토리고 이 User테이블의 pk는 Integer다.
     // select * from user where username = 1?; 가 실행
     Optional<User> findByUsername(String username);
